refactor: Migrate dating app in a Gradle module.
- Create :animations:lottie module
- Create :animations:canvas module
- Create :components:comingsoon module
- Create :demos:datingapp module

Closes #61
