+++ animations/lottie/src/main/assets/cryptoload.json
