+++ animations/lottie/src/main/assets/food.json
