+++ animations/lottie/src/main/assets/loader.json
