+++ animations/lottie/src/main/assets/working.json
