+++ animations/lottie/src/main/java/com/guru/composecookbook/lottie/LottieWorkingLoadingView.kt
@@ -1,0 +1,19 @@
+package com.guru.composecookbook.lottie
+
+import android.content.Context
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun LottieWorkingLoadingView(context: Context) {
+    LottieLoadingView(
+        context = context,
+        file = "working.json",
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(250.dp)
+    )
+}
