+++ app/src/main/java/com/guru/composecookbook/ui/animation/TransitionsAnimations.kt
@@ -3,8 +3,9 @@ package com.guru.composecookbook.ui.animation
 import MyAnimationState
 import androidx.compose.animation.animateColor
 import androidx.compose.animation.core.*
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
 import androidx.compose.material.Divider
 import androidx.compose.material.FloatingActionButton
 import androidx.compose.material.Icon
@@ -12,12 +13,11 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.PlayCircleFilled
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.guru.composecookbook.canvas.MultiStateAnimationCircleFilledCanvas
+import com.guru.composecookbook.canvas.MultiStateAnimationCircleStrokeCanvas
 import com.guru.composecookbook.theme.green200
 import com.guru.composecookbook.theme.green500
 import com.guru.composecookbook.theme.orange
@@ -115,37 +115,7 @@ fun MultiStateInfiniteTransition() {
 @Composable
 fun FloatMultiStateAnimationCircleStrokeCanvas() {
     SubtitleText(subtitle = "Canvas Circle stroke with transition")
-    val transition = rememberInfiniteTransition()
-
-    val animatedFloat by transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 360f,
-        animationSpec = infiniteRepeatable(tween(800), RepeatMode.Restart)
-    )
-
-    val stroke = Stroke(8f)
-    Canvas(
-        modifier = Modifier
-            .padding(16.dp)
-            .size(100.dp)
-    ) {
-        val diameter = size.minDimension
-        val radius = diameter / 2f
-        val insideRadius = radius - stroke.width
-        val topLeftOffset = Offset(10f, 10f)
-        val size = Size(insideRadius * 2, insideRadius * 2)
-        var rotationAngle = animatedFloat - 50f
-        drawArc(
-            color = green500,
-            startAngle = rotationAngle,
-            sweepAngle = 150f,
-            topLeft = topLeftOffset,
-            size = size,
-            useCenter = false,
-            style = stroke,
-        )
-        rotationAngle += 40
-    }
+    MultiStateAnimationCircleStrokeCanvas()
 }
 
 @Preview
@@ -153,34 +123,6 @@ fun FloatMultiStateAnimationCircleStrokeCanvas() {
 fun FloatMultiStateAnimationCircleCanvas(color: Color = green500, radiusEnd: Float = 200f) {
     SubtitleText(subtitle = "Canvas drawing with infinite transition")
     Spacer(modifier = Modifier.height(100.dp))
-    val transition = rememberInfiniteTransition()
-    val floatAnim by transition.animateFloat(
-        initialValue = 10f,
-        targetValue = radiusEnd,
-        animationSpec = infiniteRepeatable(tween(1200), RepeatMode.Reverse)
-    )
-
-    Canvas(modifier = Modifier.padding(16.dp)) {
-        val centerOffset = Offset(
-            10f,
-            10f
-        )
-        val radius = floatAnim
-        drawCircle(
-            color = color.copy(alpha = 0.8f),
-            radius = radius,
-            center = centerOffset,
-        )
-        drawCircle(
-            color = color.copy(alpha = 0.4f),
-            radius = radius / 2,
-            center = centerOffset,
-        )
-        drawCircle(
-            color = color.copy(alpha = 0.2f),
-            radius = radius / 4,
-            center = centerOffset,
-        )
-    }
+    MultiStateAnimationCircleFilledCanvas(color, radiusEnd)
     Spacer(modifier = Modifier.height(100.dp))
 }
