--- app/src/main/java/com/guru/composecookbook/ui/demoapps/datingapp/DatingHomeViewModel.kt
@@ -1,5 +1,0 @@
-package com.guru.composecookbook.ui.demoapps.datingapp
-
-import androidx.lifecycle.ViewModel
-
-class DatingHomeViewModel : ViewModel()
