+++ app/src/main/java/com/guru/composecookbook/ui/utils/CommomComponents.kt
@@ -4,49 +4,21 @@ import FaIcons
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.*
+import androidx.compose.material.Divider
+import androidx.compose.material.LocalContentAlpha
+import androidx.compose.material.LocalContentColor
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.guru.composecookbook.theme.typography
-import com.guru.composecookbook.ui.templates.logins.LottieLoadingView
 import com.guru.fontawesomecomposelib.FaIcon
-
-@Composable
-fun ComingSoon() {
-    Surface {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(50.dp)
-        ) {
-            LottieLoadingView(context = LocalContext.current)
-            Text(
-                text = "Coming Soon",
-                style = typography.h5,
-                modifier = Modifier
-                    .padding(12.dp)
-                    .fillMaxWidth(),
-                textAlign = TextAlign.Center
-            )
-            Text(
-                text = "work in progress",
-                style = typography.subtitle2,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
 
 
 @Composable
