+++ app/src/main/res/values/strings.xml
@@ -19,11 +19,9 @@
 
     <string name="test_adbanner_id">ca-app-pub-3940256099942544/6300978111</string>
     <string name="test_admob_id">ca-app-pub-5207778306358129~3940384518</string>
-    <string name="title_activity_dating_home_activitu">DatingHome</string>
     <string name="title_activity_advance_lists">My Compose App</string>
     <string name="title_activity_custom_fling">Custom Fling</string>
     <string name="title_activity_custom_fling_settings">Custom Fling Settings</string>
-    <string name="title_activity_tiktok">Tiktok</string>
     <string name="title_activity_templates">Templates</string>
     <string name="card_holder_name">Card holder name</string>
     <string name="card_holder_number">Card number</string>
