+++ demos/datingapp/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.guru.composecookbook.datingapp">
+
+    <application>
+        <activity
+            android:name=".DatingHomeActivity"
+            android:label="@string/title_activity_dating"
+            android:theme="@style/Theme.ComposeCookBook.NoActionBar" />
+    </application>
+</manifest>
