+++ demos/datingapp/src/main/java/com/guru/composecookbook/datingapp/components/profile/ProfileScreen.kt
@@ -1,4 +1,4 @@
-package com.guru.composecookbook.ui.demoapps.datingapp.components
+package com.guru.composecookbook.datingapp.components.profile
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -7,11 +7,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.guru.composecookbook.ui.utils.ComingSoon
+import com.guru.composecookbook.comingsoon.ComingSoon
 
 @Composable
 fun ProfileScreen() {
-
     LazyColumn(
         modifier = Modifier
             .fillMaxWidth()
