+++ demos/datingapp/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="title_activity_dating">DatingHome</string>
+</resources>
