+++ settings.gradle.kts
@@ -19,3 +19,7 @@ include(":components:fab")
 include(":demos:paint")
 include(":components:colorpicker")
 include(":demos:tiktok")
+include(":demos:datingapp")
+include(":animations:lottie")
+include(":animations:canvas")
+include(":components:comingsoon")
